refactor: converted MainActivity.java to kotlin, moved MainActivity & MainActivityFragment to new packages
