+++ app/src/main/kotlin/za/foundation/praekelt/mama/app/fragment/MainActivityFragment.kt
@@ -1,10 +1,11 @@
-package za.foundation.praekelt.mama
+package za.foundation.praekelt.mama.app.fragment
 
-import android.support.v4.app.Fragment
 import android.os.Bundle
+import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import za.foundation.praekelt.mama.R
 
 /**
  * A placeholder fragment containing a simple view.
